+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v1.4.0 - 2023-11-20
+### Added
+- `get_conjuncts(p: HplPredicate | HplExpression)` function to `hpl.rewrite` module.
+- `get_disjuncts(p: HplPredicate | HplExpression)` function to `hpl.rewrite` module.
+
+### Changed
+- Small improvement in `hpl.rewrite.simplify`.
+
 ## v1.3.0 - 2023-10-04
 ### Added
 - `split_and(p: HplPredicate | HplExpression)` function to `hpl.rewrite` module.